The Seperation of Old and New
