+++ oid4vc-manager/tests/oid4vp/mod.rs
@@ -1,2 +1,2 @@
-pub mod dcql_implicit;
 pub mod implicit;
+pub mod old_implicit;
