--- oid4vp/src/token/dcql_vp_token_builder.rs
@@ -1,116 +1,0 @@
-use crate::dcql::dcql_query::{CredentialId, CredentialQuery, DcqlQuery};
-use crate::token::dcql_vp_token::{DcqlQueryVpToken, PresentationFormat};
-use anyhow::Result;
-use std::collections::HashMap;
-use validator::{Validate, ValidationError, ValidationErrors};
-
-#[derive(Default, Validate)]
-pub struct DcqlVpTokenBuilder {
-    presentations: HashMap<CredentialId, Vec<PresentationFormat>>,
-    dcql_query: Option<DcqlQuery>,
-}
-
-impl Default for DcqlQueryVpToken {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl DcqlVpTokenBuilder {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn builder_dcql_query(dcql_query: DcqlQuery) -> Self {
-        Self {
-            presentations: HashMap::new(),
-            dcql_query: Some(dcql_query),
-        }
-    }
-
-    pub fn add_presentation(mut self, credential_id: CredentialId, presentation: PresentationFormat) -> Self {
-        self.presentations.entry(credential_id).or_default().push(presentation);
-        self
-    }
-
-    /// for multiple presentations for the same credential_id
-    pub fn add_presentations(mut self, credential_id: CredentialId, presentations: Vec<PresentationFormat>) -> Self {
-        self.presentations.insert(credential_id, presentations);
-        self
-    }
-
-    pub fn build(self) -> Result<DcqlQueryVpToken, ValidationErrors> {
-        self.validate()?;
-        if let Some(ref dcql_query) = self.dcql_query {
-            dcql_query.validate()?;
-            self.validate_against_dcql(dcql_query)?;
-        }
-
-        Ok(DcqlQueryVpToken {
-            presentations: self.presentations,
-        })
-    }
-
-    fn validate_against_dcql(&self, dcql_query: &DcqlQuery) -> Result<(), ValidationErrors> {
-        let mut errors = ValidationErrors::new();
-
-        let credential_queries: HashMap<CredentialId, &CredentialQuery> =
-            dcql_query.credentials.iter().map(|cq| (cq.id.clone(), cq)).collect();
-
-        // Check to see if all required credentials are present.
-        for credential_query in &dcql_query.credentials {
-            if !self.presentations.contains_key(&credential_query.id) {
-                let mut error = ValidationError::new("missing_required_credential");
-                error.message = Some(
-                    format!(
-                        "Required credential '{}' is missing from presentations",
-                        credential_query.id
-                    )
-                    .into(),
-                );
-                errors.add("presentations", error);
-            }
-        }
-
-        // Make sure there are no extra or unnecessary presentations
-        for credential_id in self.presentations.keys() {
-            if !credential_queries.contains_key(credential_id) {
-                let mut error = ValidationError::new("invalid_credential_id");
-                error.message = Some(
-                    format!(
-                        "Presentation provided for '{}' which was not requested for.",
-                        credential_id
-                    )
-                    .into(),
-                );
-                errors.add("presentations", error);
-            }
-        }
-
-        // Check multiple parameter
-        for (credential_id, presentations) in &self.presentations {
-            if let Some(credential_query) = credential_queries.get(credential_id) {
-                let allows_multiple = credential_query.multiple.unwrap_or(false);
-
-                if !allows_multiple && presentations.len() > 1 {
-                    let mut error = ValidationError::new("multiple_not_allowed");
-                    error.message = Some(
-                        format!(
-                            "Credential '{}' does not allow multiple presentations, but {} provided",
-                            credential_id,
-                            presentations.len()
-                        )
-                        .into(),
-                    );
-                    errors.add("presentations", error);
-                }
-            }
-        }
-
-        if errors.is_empty() {
-            Ok(())
-        } else {
-            Err(errors)
-        }
-    }
-}
