+++ oid4vp/src/token/mod.rs
@@ -1,4 +1,4 @@
-pub mod dcql_vp_token;
-pub mod dcql_vp_token_builder;
+pub mod old_vp_token;
+pub mod old_vp_token_builder;
 pub mod vp_token;
 pub mod vp_token_builder;
